Restructuring the initial setup code
