+++ pubspec.yaml
@@ -18,6 +18,7 @@ dependencies:
   flutter_svg: ^2.0.5
   form_builder_validators: ^9.0.0
   http: ^1.1.2
+  intl: ^0.19.0
   jwt_decoder: ^2.0.1
   loading_indicator: ^3.1.1
   markdown_editor_plus: ^0.2.13
